Created ShoppingListLogo.kt and refactored portrait screens
