+++ composeApp/src/commonMain/kotlin/io/dimasla4ee/shoppinglist/feature/authorization/ui/AuthorizationScreen.kt
@@ -1,11 +1,10 @@
 package io.dimasla4ee.shoppinglist.feature.authorization.ui
 
-import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.defaultMinSize
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
@@ -13,19 +12,17 @@ import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.MediumTopAppBar
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import io.dimasla4ee.shoppinglist.app.ui.theme.AppDimensions
 import io.dimasla4ee.shoppinglist.app.ui.theme.ShoppingListTheme
+import io.dimasla4ee.shoppinglist.core.presentation.components.ShoppingListLogo
 import io.dimasla4ee.shoppinglist.core.presentation.components.topbar.AppTopBarDefaults
 import io.dimasla4ee.shoppinglist.feature.authorization.presentation.recover_password.RecoverPasswordState
 import io.dimasla4ee.shoppinglist.feature.authorization.presentation.register.RegisterState
@@ -33,7 +30,6 @@ import io.dimasla4ee.shoppinglist.feature.authorization.presentation.sign_in.Sig
 import io.dimasla4ee.shoppinglist.feature.authorization.ui.recover_password.RecoverPasswordContent
 import io.dimasla4ee.shoppinglist.feature.authorization.ui.register.RegisterContent
 import io.dimasla4ee.shoppinglist.feature.authorization.ui.sign_in.SignInContent
-import io.dimasla4ee.shoppinglist.feature.welcome_screen.ui.createWelcomeLogo
 import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.resources.stringResource
 import shoppinglist.composeapp.generated.resources.Res
@@ -47,23 +43,13 @@ fun AuthorizationScreen(
     modifier: Modifier = Modifier,
     content: @Composable () -> Unit
 ) {
-    val (annotatedString, inlineContentMap) = createWelcomeLogo()
     val scrollState = rememberScrollState()
 
     Scaffold(
         modifier = modifier,
         topBar = {
-            MediumTopAppBar(
-                title = {
-                    Text(
-                        modifier = Modifier.fillMaxWidth(),
-                        inlineContent = inlineContentMap,
-                        text = annotatedString,
-                        style = MaterialTheme.typography.titleLargeEmphasized,
-                        color = MaterialTheme.colorScheme.onBackground,
-                        textAlign = TextAlign.Center
-                    )
-                },
+            TopAppBar(
+                title = {},
                 navigationIcon = {
                     IconButton(onClick = onBackClick) {
                         Icon(
@@ -84,13 +70,19 @@ fun AuthorizationScreen(
         ) {
             Column(
                 modifier = Modifier
-                    .fillMaxWidth()
+                    .fillMaxSize()
                     .defaultMinSize(minHeight = maxHeight)
                     .verticalScroll(scrollState),
-                verticalArrangement = Arrangement.Center,
                 horizontalAlignment = Alignment.CenterHorizontally
             ) {
-                content()
+                ShoppingListLogo {}
+
+                Box(
+                    modifier = Modifier.weight(1F),
+                    contentAlignment = Alignment.Center
+                ) {
+                    content()
+                }
             }
         }
     }
