+++ composeApp/src/commonMain/kotlin/io/dimasla4ee/shoppinglist/feature/products_screen/ui/ItemListPlaceholder.kt
@@ -1,6 +1,7 @@
 package io.dimasla4ee.shoppinglist.feature.products_screen.ui
 
 import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
@@ -12,14 +13,13 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
+import io.dimasla4ee.shoppinglist.app.ui.theme.AppDimensions
 import io.dimasla4ee.shoppinglist.app.ui.theme.LocalAppPlaceholders
 import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.resources.stringResource
 import shoppinglist.composeapp.generated.resources.Res
 import shoppinglist.composeapp.generated.resources.add_items_hint
 import shoppinglist.composeapp.generated.resources.empty_list_message
-import shoppinglist.composeapp.generated.resources.img_product_list
 
 @Composable
 fun ItemListPlaceholder(
@@ -27,33 +27,33 @@ fun ItemListPlaceholder(
 ) {
     // Основной контент экрана
     Column(
-        modifier = modifier.fillMaxSize(),
-        horizontalAlignment = Alignment.CenterHorizontally
+        modifier = modifier
+            .fillMaxSize()
+            .padding(horizontal = AppDimensions.paddingLarge),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.Center
     ) {
         Image(
             painter = painterResource(LocalAppPlaceholders.current.imgProductList),
             contentDescription = null
         )
 
-        Spacer(modifier = Modifier.height(48.dp))
+        Spacer(modifier = Modifier.height(AppDimensions.spacerVeryBig))
 
         Text(
-            modifier = Modifier.padding(horizontal = 44.dp),
+            textAlign = TextAlign.Center,
             text = stringResource(Res.string.empty_list_message),
             style = MaterialTheme.typography.labelLarge,
             color = MaterialTheme.colorScheme.onBackground
         )
 
-        Spacer(modifier = Modifier.height(8.dp))
+        Spacer(modifier = Modifier.height(AppDimensions.spacerSmall))
 
         Text(
-            modifier = Modifier.padding(horizontal = 44.dp),
             textAlign = TextAlign.Center,
             text = stringResource(Res.string.add_items_hint),
             style = MaterialTheme.typography.bodyMedium,
             color = MaterialTheme.colorScheme.onBackground
         )
-
-        Spacer(modifier = Modifier.weight(1f))
     }
 }
